requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
idna==3.4
kiwisolver==1.4.4
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.2
